+++ lib/lousy/widget/zoom.lua
@@ -10,6 +10,7 @@ local webview = require("webview")
 local lousy = require("lousy")
 local theme = lousy.theme.get()
 local wc = require("lousy.widget.common")
+local globals = require("globals")
 
 local _M = {}
 
@@ -22,7 +23,7 @@ _M.format = "[zoom:%d%%]"
 local widgets = {
     update = function (w, zoom)
         local zl = w.view.zoom_level
-        if zl == 1.0 then
+        if zl == (globals.default_zoom_level or 100)/100 then
             zoom:hide()
         else
             zoom:show()